+++ src/utils/JsonParser.sol
@@ -3,7 +3,6 @@
 pragma solidity ^0.8.24;
 
 library JsonParser {
-
     /**
      * @dev Extracts the value of a given key from a JSON string, supports nested keys.
      * @param json The JSON string to parse.
@@ -15,11 +14,11 @@ library JsonParser {
         bytes memory keyBytes = bytes(key);
 
         // Find the starting position of the key
-        for (uint i = 0; i < jsonBytes.length; i++) {
+        for (uint256 i = 0; i < jsonBytes.length; i++) {
             if (isMatch(jsonBytes, keyBytes, i)) {
                 // Skip over the `":"` character
-                uint start = i + keyBytes.length + 3;
-                uint end = start;
+                uint256 start = i + keyBytes.length + 3;
+                uint256 end = start;
 
                 // Search for the closing quote of the value
                 while (end < jsonBytes.length && jsonBytes[end] != '"') {
@@ -27,7 +26,7 @@ library JsonParser {
                 }
 
                 bytes memory valueBytes = new bytes(end - start);
-                for (uint j = start; j < end; j++) {
+                for (uint256 j = start; j < end; j++) {
                     valueBytes[j - start] = jsonBytes[j];
                 }
 
@@ -45,12 +44,12 @@ library JsonParser {
      * @param start The starting index to compare.
      * @return True if the key matches, false otherwise.
      */
-    function isMatch(bytes memory jsonBytes, bytes memory keyBytes, uint start) internal pure returns (bool) {
+    function isMatch(bytes memory jsonBytes, bytes memory keyBytes, uint256 start) internal pure returns (bool) {
         if (start + keyBytes.length >= jsonBytes.length) {
             return false;
         }
 
-        for (uint i = 0; i < keyBytes.length; i++) {
+        for (uint256 i = 0; i < keyBytes.length; i++) {
             if (jsonBytes[start + i] != keyBytes[i]) {
                 return false;
             }
